pyproject.toml
[project]
name = "pkb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Jens Nilsson", email = "jenszza@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.0",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-huggingface>=0.3.0",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.2.20,<0.3",
    "llama-cpp-python>=0.3.9",
    "sentence-transformers>=4.1.0",
]

[project.scripts]
pkb = "pkb:main"

[tool.ruff]
line-length = 88

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["pytest>=8.4.0", "ruff>=0.11.13"]
